fix: stop split-focus flicker from overlapping focus retry loops
focusSplitPane spawned an independent 12x30ms makeFirstResponder retry loop per
call with no cancellation, so two calls with opposite targets (a tmux-style zoom
keybind firing `session focus` back-to-back) ran concurrently and ping-ponged
first responder between the panes for ~400ms - splitFocused flipped ~13 times and
both surfaces redrew on each flip, the visible flicker. Pre-existing (since
<=0.11.0), intermittent, and only from rapid opposite focus ops, never plain cmd+d.

a focusGeneration token now supersedes older loops: a fresh call bumps it, and an
in-flight loop stops on its next tick once its captured generation is stale. The
surviving loop still re-asserts through the split-materialize / reparent churn
(re-asserts are no-ops once the target holds first responder), keeping the retry's
original purpose.

focusSplitPane moves to a new AppActions+Focus.swift so AppActions.swift stays
under the 1000-line cap; focusGeneration stays in the main class (extensions can't
hold stored properties), with frontmostQuickTerminal and terminalZoomActive(for:)
widened to internal so the extension can reach them.
